GED/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
dnspython==2.6.1
pymongo==4.7.2
sqlparse==0.5.0
typing_extensions==4.12.0
